+++ frontend/apps/investment-calculator-app/src/components/Header/Header.jsx
@@ -2,8 +2,8 @@ const logo =
     'https://s3.eu-central-1.amazonaws.com/assets.reactjs-component-lib.eu/images/investment-calculator-logo-original.webp';
 
 import classes from './Header.module.scss';
-import '@prj--reactjs-component-lib/generic-lsg/lib/atoms/heading.css';
-import { Heading } from '@prj--reactjs-component-lib/generic-lib';
+import '@prj--reactjs-application-library/generic-lsg/lib/atoms/heading.css';
+import { Heading } from '@prj--reactjs-application-library/generic-lib';
 
 const Header = () => {
     return (