+++ frontend/apps/login-form-app/src/App.jsx
@@ -7,7 +7,7 @@ import AuthContext from './store/auth-context';
 import { loginForm } from './LoginForm.module.scss';
 import { AuthContextProvider } from './store/auth-context';
 
-import '@prj--reactjs-component-lib/generic-lsg/lib/global.css';
+import '@prj--reactjs-application-library/generic-lsg/lib/global.css';
 
 function LoginFormCmp() {
     const ctx = useContext(AuthContext);