+++ frontend/apps/react-essentials-app/src/components/organisms/Examples.jsx
@@ -1,7 +1,7 @@
-import '@prj--reactjs-component-lib/generic-lsg/lib/atoms/heading.css';
-import '@prj--reactjs-component-lib/generic-lsg/lib/atoms/paragraph.css';
+import '@prj--reactjs-application-library/generic-lsg/lib/atoms/heading.css';
+import '@prj--reactjs-application-library/generic-lsg/lib/atoms/paragraph.css';
 
-import { Heading, Paragraph } from '@prj--reactjs-component-lib/generic-lib';
+import { Heading, Paragraph } from '@prj--reactjs-application-library/generic-lib';
 import Tabs from '../molecules/Tabs';
 import Section from '../molecules/Section';
 import TabButton from '../atoms/TabButton';