+++ frontend/apps/investment-calculator-app/CHANGELOG.md
@@ -3,39 +3,72 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.7.0 (2023-11-08)
+
+### Bug Fixes
+
+-   **rename:** rename ([fa982ac](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fa982ac05c386e2f9563611d754b73bc08b54894))
+
+### Features
+
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9436ecafd5addb266153737a53f95922733b9a63))
+-   cleanup ([fe007c8](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fe007c8c6119a0a7f875b185dfd586117dc24fbe))
+-   formats ([d81e33d](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/d81e33db14d484d87c5c29b249d9d2cbd3a13560))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   migrated investment calculator ([7a4751b](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/7a4751bc21a85194e1caa99b07ad693834e1de35))
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+-   text-app depends on geenric-lib ([1ad92ea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/1ad92eaae9a45363ffd4876bf89218c87f798de0))
+
+<<<<<<< Updated upstream
+
 # [0.6.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/investment-calculator-app@0.5.0...@prj--reactjs-component-lib/investment-calculator-app@0.6.0) (2023-11-08)
 
 ### Features
 
--   cleanup ([fe007c8](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/fe007c8c6119a0a7f875b185dfd586117dc24fbe))
+-   # cleanup ([fe007c8](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/fe007c8c6119a0a7f875b185dfd586117dc24fbe))
+    <<<<<<< Updated upstream
+    > > > > > > > Stashed changes
 
 # [0.5.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/investment-calculator-app@0.4.0...@prj--reactjs-component-lib/investment-calculator-app@0.5.0) (2023-10-31)
 
-### Features
+=======
 
--   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/9436ecafd5addb266153737a53f95922733b9a63))
-
-# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/investment-calculator-app@0.3.0...@prj--reactjs-component-lib/investment-calculator-app@0.4.0) (2023-10-30)
+# [0.6.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/investment-calculator-app@0.5.0...@prj--reactjs-application-library/investment-calculator-app@0.6.0) (2023-11-08)
 
 ### Features
 
--   formats ([d81e33d](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/d81e33db14d484d87c5c29b249d9d2cbd3a13560))
+-   cleanup ([fe007c8](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fe007c8c6119a0a7f875b185dfd586117dc24fbe))
 
-# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/investment-calculator-app@0.2.1...@prj--reactjs-component-lib/investment-calculator-app@0.3.0) (2023-10-30)
+# [0.5.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/investment-calculator-app@0.4.0...@prj--reactjs-application-library/investment-calculator-app@0.5.0) (2023-10-31)
+
+> > > > > > > Stashed changes
 
 ### Features
 
--   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
--   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9436ecafd5addb266153737a53f95922733b9a63))
 
-## [0.2.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/investment-calculator-app@0.2.0...@prj--reactjs-component-lib/investment-calculator-app@0.2.1) (2023-10-30)
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/investment-calculator-app@0.3.0...@prj--reactjs-application-library/investment-calculator-app@0.4.0) (2023-10-30)
 
-**Note:** Version bump only for package @prj--reactjs-component-lib/investment-calculator-app
+### Features
+
+-   formats ([d81e33d](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/d81e33db14d484d87c5c29b249d9d2cbd3a13560))
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/investment-calculator-app@0.2.1...@prj--reactjs-application-library/investment-calculator-app@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+
+## [0.2.1](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/investment-calculator-app@0.2.0...@prj--reactjs-application-library/investment-calculator-app@0.2.1) (2023-10-30)
+
+**Note:** Version bump only for package @prj--reactjs-application-library/investment-calculator-app
 
 # 0.2.0 (2023-10-26)
 
 ### Features
 
--   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
--   migrated investment calculator ([7a4751b](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/7a4751bc21a85194e1caa99b07ad693834e1de35))
--   text-app depends on geenric-lib ([1ad92ea](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/1ad92eaae9a45363ffd4876bf89218c87f798de0))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   migrated investment calculator ([7a4751b](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/7a4751bc21a85194e1caa99b07ad693834e1de35))
+-   text-app depends on geenric-lib ([1ad92ea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/1ad92eaae9a45363ffd4876bf89218c87f798de0))