+++ frontend/apps/login-form-app/CHANGELOG.md
@@ -3,37 +3,69 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.6.0 (2023-11-08)
+
+### Bug Fixes
+
+-   **rename:** rename ([fa982ac](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fa982ac05c386e2f9563611d754b73bc08b54894))
+
+### Features
+
+-   added login form ([67481d4](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/67481d4f1abecb1a382ffa5f820c7ff8acc25295))
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9436ecafd5addb266153737a53f95922733b9a63))
+-   cleanup ([f02ce8f](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f02ce8fc7ff57b5b5d0514d0744105b0f887e0ed))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+-   text-app depends on geenric-lib ([1ad92ea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/1ad92eaae9a45363ffd4876bf89218c87f798de0))
+
+<<<<<<< Updated upstream
+
 # [0.5.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/login-form-app@0.4.0...@prj--reactjs-component-lib/login-form-app@0.5.0) (2023-11-08)
 
 ### Features
 
--   cleanup ([f02ce8f](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f02ce8fc7ff57b5b5d0514d0744105b0f887e0ed))
+-   # cleanup ([f02ce8f](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f02ce8fc7ff57b5b5d0514d0744105b0f887e0ed))
+    <<<<<<< Updated upstream
+    > > > > > > > Stashed changes
 
 # [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/login-form-app@0.3.1...@prj--reactjs-component-lib/login-form-app@0.4.0) (2023-10-31)
 
-### Features
+=======
 
--   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/9436ecafd5addb266153737a53f95922733b9a63))
-
-## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/login-form-app@0.3.0...@prj--reactjs-component-lib/login-form-app@0.3.1) (2023-10-30)
-
-**Note:** Version bump only for package @prj--reactjs-component-lib/login-form-app
-
-# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/login-form-app@0.2.1...@prj--reactjs-component-lib/login-form-app@0.3.0) (2023-10-30)
+# [0.5.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/login-form-app@0.4.0...@prj--reactjs-application-library/login-form-app@0.5.0) (2023-11-08)
 
 ### Features
 
--   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
--   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+-   cleanup ([f02ce8f](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f02ce8fc7ff57b5b5d0514d0744105b0f887e0ed))
 
-## [0.2.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/login-form-app@0.2.0...@prj--reactjs-component-lib/login-form-app@0.2.1) (2023-10-30)
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/login-form-app@0.3.1...@prj--reactjs-application-library/login-form-app@0.4.0) (2023-10-31)
 
-**Note:** Version bump only for package @prj--reactjs-component-lib/login-form-app
+> > > > > > > Stashed changes
+
+### Features
+
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9436ecafd5addb266153737a53f95922733b9a63))
+
+## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/login-form-app@0.3.0...@prj--reactjs-application-library/login-form-app@0.3.1) (2023-10-30)
+
+**Note:** Version bump only for package @prj--reactjs-application-library/login-form-app
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/login-form-app@0.2.1...@prj--reactjs-application-library/login-form-app@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+
+## [0.2.1](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/login-form-app@0.2.0...@prj--reactjs-application-library/login-form-app@0.2.1) (2023-10-30)
+
+**Note:** Version bump only for package @prj--reactjs-application-library/login-form-app
 
 # 0.2.0 (2023-10-26)
 
 ### Features
 
--   added login form ([67481d4](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/67481d4f1abecb1a382ffa5f820c7ff8acc25295))
--   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
--   text-app depends on geenric-lib ([1ad92ea](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/1ad92eaae9a45363ffd4876bf89218c87f798de0))
+-   added login form ([67481d4](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/67481d4f1abecb1a382ffa5f820c7ff8acc25295))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   text-app depends on geenric-lib ([1ad92ea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/1ad92eaae9a45363ffd4876bf89218c87f798de0))