+++ frontend/apps/lottery-app/CHANGELOG.md
@@ -3,32 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.5.0 (2023-11-08)
+
+### Bug Fixes
+
+-   **rename:** rename ([fa982ac](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fa982ac05c386e2f9563611d754b73bc08b54894))
+
+### Features
+
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9436ecafd5addb266153737a53f95922733b9a63))
+-   **lottery-app:** init ([8549f66](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/8549f66544931aa542c2765e04ec1dec0007aaad))
+-   **lottery:** initial implementation ([3254cb4](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/3254cb4d5806aa65f25342a41184c43c01c8cc7f))
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+-   small ([017fed7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/017fed7446c5b761e8bd32e1e9ca77aec2ed590f))
+
+<<<<<<< Updated upstream
+
 ## [0.4.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/lottery-app@0.4.0...@prj--reactjs-component-lib/lottery-app@0.4.1) (2023-11-08)
 
-**Note:** Version bump only for package @prj--reactjs-component-lib/lottery-app
+# **Note:** Version bump only for package @prj--reactjs-component-lib/lottery-app
+
+<<<<<<< Updated upstream
+
+> > > > > > > Stashed changes
 
 # [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/lottery-app@0.3.0...@prj--reactjs-component-lib/lottery-app@0.4.0) (2023-10-31)
 
-### Features
+=======
 
--   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/9436ecafd5addb266153737a53f95922733b9a63))
--   small ([017fed7](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/017fed7446c5b761e8bd32e1e9ca77aec2ed590f))
+## [0.4.1](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/lottery-app@0.4.0...@prj--reactjs-application-library/lottery-app@0.4.1) (2023-11-08)
 
-# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/lottery-app@0.2.0...@prj--reactjs-component-lib/lottery-app@0.3.0) (2023-10-30)
+**Note:** Version bump only for package @prj--reactjs-application-library/lottery-app
 
-### Features
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/lottery-app@0.3.0...@prj--reactjs-application-library/lottery-app@0.4.0) (2023-10-31)
 
--   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
--   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
-
-# [0.2.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/lottery-app@0.1.0...@prj--reactjs-component-lib/lottery-app@0.2.0) (2023-10-29)
+> > > > > > > Stashed changes
 
 ### Features
 
--   **lottery:** initial implementation ([3254cb4](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/3254cb4d5806aa65f25342a41184c43c01c8cc7f))
+-   **base path:** fixed ([9436eca](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9436ecafd5addb266153737a53f95922733b9a63))
+-   small ([017fed7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/017fed7446c5b761e8bd32e1e9ca77aec2ed590f))
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/lottery-app@0.2.0...@prj--reactjs-application-library/lottery-app@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package:** added package scripts to apps ([15349d0](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/15349d0e3d3eac4222a99a42b28d4d67b764557f))
+
+# [0.2.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/lottery-app@0.1.0...@prj--reactjs-application-library/lottery-app@0.2.0) (2023-10-29)
+
+### Features
+
+-   **lottery:** initial implementation ([3254cb4](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/3254cb4d5806aa65f25342a41184c43c01c8cc7f))
 
 # 0.1.0 (2023-10-28)
 
 ### Features
 
--   **lottery-app:** init ([8549f66](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/8549f66544931aa542c2765e04ec1dec0007aaad))
+-   **lottery-app:** init ([8549f66](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/8549f66544931aa542c2765e04ec1dec0007aaad))