+++ frontend/design-systems/generic-lsg/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.5.0 (2023-11-08)
+
+### Bug Fixes
+
+-   **rename:** rename ([fa982ac](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fa982ac05c386e2f9563611d754b73bc08b54894))
+-   styles ([26dc969](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/26dc969af330dda8d62faf01cd1de9e6ece542fd))
+
+### Features
+
+-   added counter compont ([e30faea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e30faeaa5dcfcb048205dcd92b5ef57000c47acc))
+-   added login form ([67481d4](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/67481d4f1abecb1a382ffa5f820c7ff8acc25295))
+-   added rating component ([871e393](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/871e393f0f2132d22529487a60eaa1bb9dd03432))
+-   added tests ([545ee7e](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/545ee7ed79bb6c88ebbb863c143fd949c548bda4))
+-   added user age form ([3dc973f](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/3dc973f8ea79a06180adead51d090808e4db0f74))
+-   chromatic updates ([73c7300](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/73c730043a99e4c89ce32ff79b614fde9b945146))
+-   cleanup ([fe007c8](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fe007c8c6119a0a7f875b185dfd586117dc24fbe))
+-   cleanup ([e0b8abc](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e0b8abce990c1f5cfd6cf7611a534da53fea446b))
+-   cleanup ([054e587](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/054e587ee273e2c7254472f3d072ed1f299c22c4))
+-   cleanup ([bdc0a92](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/bdc0a92044bd5b3691ec5f56c9ea2f86e3a7802e))
+-   cleanup fix ([120fdea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/120fdeab392739c535161d5a84eea9c3fe337961))
+-   **icon:** star to icon ([66441c3](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/66441c346c41da6de5139894558be64060a0206c))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   migrated concept cards ([c4170d2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/c4170d2130e71d04e587acd0f9a4f1becef4d0b3))
+-   migrated expense tracker ([657a6be](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/657a6bef2fb694fa88de672cb984d293dfe4847c))
+-   migrated investment calculator ([7a4751b](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/7a4751bc21a85194e1caa99b07ad693834e1de35))
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **paragraph:** from text to paragraph ([43952f3](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/43952f31ec26c6e1fefd16513e1053fe5b857345))
+-   **react essentials app:** course implementation ([bb97dc5](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/bb97dc5c0793c9e19bc5b01da1733f97d91988b4))
+-   **react-essentials:** added core concepts ([c39d7ab](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/c39d7ab1412af79df1e6602da05f4b08cd1e3e54))
+-   some small changes ([5a96b20](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/5a96b201034042e2f71b4cd0abaa245a2df90822))
+-   workflow implementation ([dbae96d](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/dbae96dfe108f8a2638051cb727efc6b86b606d4))
+
+<<<<<<< Updated upstream
+
 # [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/generic-lsg@0.3.1...@prj--reactjs-component-lib/generic-lsg@0.4.0) (2023-11-08)
 
 ### Bug Fixes
@@ -19,31 +53,57 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 -   **icon:** star to icon ([66441c3](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/66441c346c41da6de5139894558be64060a0206c))
 -   **paragraph:** from text to paragraph ([43952f3](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/43952f31ec26c6e1fefd16513e1053fe5b857345))
 -   **react essentials app:** course implementation ([bb97dc5](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/bb97dc5c0793c9e19bc5b01da1733f97d91988b4))
--   **react-essentials:** added core concepts ([c39d7ab](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/c39d7ab1412af79df1e6602da05f4b08cd1e3e54))
+-   # **react-essentials:** added core concepts ([c39d7ab](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/c39d7ab1412af79df1e6602da05f4b08cd1e3e54))
+    <<<<<<< Updated upstream
+    > > > > > > > Stashed changes
 
 ## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/generic-lsg@0.3.0...@prj--reactjs-component-lib/generic-lsg@0.3.1) (2023-10-31)
 
-**Note:** Version bump only for package @prj--reactjs-component-lib/generic-lsg
+=======
 
-# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/generic-lsg@0.2.0...@prj--reactjs-component-lib/generic-lsg@0.3.0) (2023-10-30)
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/generic-lsg@0.3.1...@prj--reactjs-application-library/generic-lsg@0.4.0) (2023-11-08)
+
+### Bug Fixes
+
+-   styles ([26dc969](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/26dc969af330dda8d62faf01cd1de9e6ece542fd))
 
 ### Features
 
--   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   cleanup ([fe007c8](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fe007c8c6119a0a7f875b185dfd586117dc24fbe))
+-   cleanup ([e0b8abc](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e0b8abce990c1f5cfd6cf7611a534da53fea446b))
+-   cleanup ([054e587](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/054e587ee273e2c7254472f3d072ed1f299c22c4))
+-   cleanup ([bdc0a92](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/bdc0a92044bd5b3691ec5f56c9ea2f86e3a7802e))
+-   cleanup fix ([120fdea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/120fdeab392739c535161d5a84eea9c3fe337961))
+-   **icon:** star to icon ([66441c3](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/66441c346c41da6de5139894558be64060a0206c))
+-   **paragraph:** from text to paragraph ([43952f3](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/43952f31ec26c6e1fefd16513e1053fe5b857345))
+-   **react essentials app:** course implementation ([bb97dc5](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/bb97dc5c0793c9e19bc5b01da1733f97d91988b4))
+-   **react-essentials:** added core concepts ([c39d7ab](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/c39d7ab1412af79df1e6602da05f4b08cd1e3e54))
+
+## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/generic-lsg@0.3.0...@prj--reactjs-application-library/generic-lsg@0.3.1) (2023-10-31)
+
+> > > > > > > Stashed changes
+
+**Note:** Version bump only for package @prj--reactjs-application-library/generic-lsg
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/generic-lsg@0.2.0...@prj--reactjs-application-library/generic-lsg@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
 
 # 0.2.0 (2023-10-26)
 
 ### Features
 
--   added counter compont ([e30faea](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/e30faeaa5dcfcb048205dcd92b5ef57000c47acc))
--   added login form ([67481d4](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/67481d4f1abecb1a382ffa5f820c7ff8acc25295))
--   added rating component ([871e393](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/871e393f0f2132d22529487a60eaa1bb9dd03432))
--   added tests ([545ee7e](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/545ee7ed79bb6c88ebbb863c143fd949c548bda4))
--   added user age form ([3dc973f](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/3dc973f8ea79a06180adead51d090808e4db0f74))
--   chromatic updates ([73c7300](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/73c730043a99e4c89ce32ff79b614fde9b945146))
--   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
--   migrated concept cards ([c4170d2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/c4170d2130e71d04e587acd0f9a4f1becef4d0b3))
--   migrated expense tracker ([657a6be](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/657a6bef2fb694fa88de672cb984d293dfe4847c))
--   migrated investment calculator ([7a4751b](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/7a4751bc21a85194e1caa99b07ad693834e1de35))
--   some small changes ([5a96b20](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/5a96b201034042e2f71b4cd0abaa245a2df90822))
--   workflow implementation ([dbae96d](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/dbae96dfe108f8a2638051cb727efc6b86b606d4))
+-   added counter compont ([e30faea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e30faeaa5dcfcb048205dcd92b5ef57000c47acc))
+-   added login form ([67481d4](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/67481d4f1abecb1a382ffa5f820c7ff8acc25295))
+-   added rating component ([871e393](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/871e393f0f2132d22529487a60eaa1bb9dd03432))
+-   added tests ([545ee7e](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/545ee7ed79bb6c88ebbb863c143fd949c548bda4))
+-   added user age form ([3dc973f](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/3dc973f8ea79a06180adead51d090808e4db0f74))
+-   chromatic updates ([73c7300](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/73c730043a99e4c89ce32ff79b614fde9b945146))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   migrated concept cards ([c4170d2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/c4170d2130e71d04e587acd0f9a4f1becef4d0b3))
+-   migrated expense tracker ([657a6be](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/657a6bef2fb694fa88de672cb984d293dfe4847c))
+-   migrated investment calculator ([7a4751b](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/7a4751bc21a85194e1caa99b07ad693834e1de35))
+-   some small changes ([5a96b20](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/5a96b201034042e2f71b4cd0abaa245a2df90822))
+-   workflow implementation ([dbae96d](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/dbae96dfe108f8a2638051cb727efc6b86b606d4))