+++ frontend/design-systems/storybook/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.6.0 (2023-11-08)
+
+### Bug Fixes
+
+-   **rename:** rename ([fa982ac](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/fa982ac05c386e2f9563611d754b73bc08b54894))
+-   styles ([26dc969](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/26dc969af330dda8d62faf01cd1de9e6ece542fd))
+
+### Features
+
+-   added counter compont ([e30faea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e30faeaa5dcfcb048205dcd92b5ef57000c47acc))
+-   added prettier and lerna ([e9a8c9b](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e9a8c9b7be9b13608b9ecbcec499ab4ec3224b45))
+-   added rating component ([871e393](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/871e393f0f2132d22529487a60eaa1bb9dd03432))
+-   added storybook ([865e6a6](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/865e6a68bc610cdab9bd90959a5a923ca6aa5e7b))
+-   added tests ([545ee7e](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/545ee7ed79bb6c88ebbb863c143fd949c548bda4))
+-   chromatic updates ([73c7300](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/73c730043a99e4c89ce32ff79b614fde9b945146))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **paragraph:** from text to paragraph ([43952f3](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/43952f31ec26c6e1fefd16513e1053fe5b857345))
+-   **storybok:** package ([9c6a666](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9c6a666f8096beff0ede671c56e272a1061baa26))
+-   **storybook:** bump version ([2eece13](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/2eece1341533894ea080bb2892d14b9089277a4e))
+
+<<<<<<< Updated upstream
+
 # [0.5.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/storybook@0.4.0...@prj--reactjs-component-lib/storybook@0.5.0) (2023-11-08)
 
 ### Bug Fixes
@@ -12,28 +35,47 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Features
 
 -   **paragraph:** from text to paragraph ([43952f3](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/43952f31ec26c6e1fefd16513e1053fe5b857345))
--   **storybook:** bump version ([2eece13](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/2eece1341533894ea080bb2892d14b9089277a4e))
+-   # **storybook:** bump version ([2eece13](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/2eece1341533894ea080bb2892d14b9089277a4e))
+    <<<<<<< Updated upstream
+    > > > > > > > Stashed changes
 
 # [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/storybook@0.3.0...@prj--reactjs-component-lib/storybook@0.4.0) (2023-10-31)
 
+=======
+
+# [0.5.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/storybook@0.4.0...@prj--reactjs-application-library/storybook@0.5.0) (2023-11-08)
+
+### Bug Fixes
+
+-   styles ([26dc969](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/26dc969af330dda8d62faf01cd1de9e6ece542fd))
+
 ### Features
 
--   **storybok:** package ([9c6a666](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/9c6a666f8096beff0ede671c56e272a1061baa26))
+-   **paragraph:** from text to paragraph ([43952f3](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/43952f31ec26c6e1fefd16513e1053fe5b857345))
+-   **storybook:** bump version ([2eece13](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/2eece1341533894ea080bb2892d14b9089277a4e))
 
-# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/storybook@0.2.0...@prj--reactjs-component-lib/storybook@0.3.0) (2023-10-30)
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/storybook@0.3.0...@prj--reactjs-application-library/storybook@0.4.0) (2023-10-31)
+
+> > > > > > > Stashed changes
 
 ### Features
 
--   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **storybok:** package ([9c6a666](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/9c6a666f8096beff0ede671c56e272a1061baa26))
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/storybook@0.2.0...@prj--reactjs-application-library/storybook@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
 
 # 0.2.0 (2023-10-26)
 
 ### Features
 
--   added counter compont ([e30faea](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/e30faeaa5dcfcb048205dcd92b5ef57000c47acc))
--   added prettier and lerna ([e9a8c9b](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/e9a8c9b7be9b13608b9ecbcec499ab4ec3224b45))
--   added rating component ([871e393](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/871e393f0f2132d22529487a60eaa1bb9dd03432))
--   added storybook ([865e6a6](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/865e6a68bc610cdab9bd90959a5a923ca6aa5e7b))
--   added tests ([545ee7e](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/545ee7ed79bb6c88ebbb863c143fd949c548bda4))
--   chromatic updates ([73c7300](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/73c730043a99e4c89ce32ff79b614fde9b945146))
--   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   added counter compont ([e30faea](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e30faeaa5dcfcb048205dcd92b5ef57000c47acc))
+-   added prettier and lerna ([e9a8c9b](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/e9a8c9b7be9b13608b9ecbcec499ab4ec3224b45))
+-   added rating component ([871e393](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/871e393f0f2132d22529487a60eaa1bb9dd03432))
+-   added storybook ([865e6a6](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/865e6a68bc610cdab9bd90959a5a923ca6aa5e7b))
+-   added tests ([545ee7e](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/545ee7ed79bb6c88ebbb863c143fd949c548bda4))
+-   chromatic updates ([73c7300](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/73c730043a99e4c89ce32ff79b614fde9b945146))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))